Side-by-side images should be the same (maximum) height
This is fixed for figure1&2 in the Sweet essay, by including 1) the new resize script and 2) the onload argument in the FIRST img tag of the group:

onload="resize(this, getElementById('fig2'))”

(See Sweet .md for example.)

This should also work for three images: onload="resize(this, getElementById('fig2')), getElementById('fig3'))”)

It appears this is the only essay this case is relevant, as the others seem to already be the same size? In any case, copying the script to another essay’s .md file and filling in the onload argument for the first of the group should work.